scripts/create_deployment_configs.py
"""
Create deployment configuration files for Azure ML
"""
import os
import yaml
from pathlib import Path
from dotenv import load_dotenv

# Load environment
project_root = Path(__file__).parent.parent
load_dotenv(project_root / ".env")


def create_deployment_configs():
    """Create YAML configuration files for Azure ML deployment."""
    
    print("Creating deployment configuration files...")
    
    # Create deployment directory
    deploy_dir = project_root / "deployment"
    deploy_dir.mkdir(exist_ok=True)
    
    # 1. Endpoint configuration
    endpoint_config = {
        "$schema": "https://azuremlschemas.azureedge.net/latest/managedOnlineEndpoint.schema.json",
        "name": "outlander-copilot-endpoint",
        "description": "Outlander Gear Co. AI Copilot - RAG-based product assistant",
        "auth_mode": "key",
        "tags": {
            "project": "outlander-copilot",
            "type": "rag-chatbot",
            "framework": "promptflow"
        }
    }
    
    endpoint_file = deploy_dir / "endpoint.yaml"
    with open(endpoint_file, 'w') as f:
        yaml.dump(endpoint_config, f, default_flow_style=False, sort_keys=False)
    print(f"✓ Created: {endpoint_file}")
    
    # 2. Deployment configuration
    deployment_config = {
        "$schema": "https://azuremlschemas.azureedge.net/latest/managedOnlineDeployment.schema.json",
        "name": "outlander-deployment-v1",
        "endpoint_name": "outlander-copilot-endpoint",
        "model": {
            "path": "./prompt_flows/outlander_copilot",
            "type": "custom_model"
        },
        "code_configuration": {
            "code": "./prompt_flows/outlander_copilot",
            "scoring_script": "flow.dag.yaml"
        },
        "environment": {
            "conda_file": "./prompt_flows/outlander_copilot/requirements.txt",
            "image": "mcr.microsoft.com/azureml/promptflow/promptflow-runtime:latest"
        },
        "instance_type": "Standard_DS3_v2",
        "instance_count": 1,
        "environment_variables": {
            "AZURE_OPENAI_ENDPOINT": os.getenv("AZURE_OPENAI_ENDPOINT"),
            "AZURE_OPENAI_API_VERSION": os.getenv("AZURE_OPENAI_API_VERSION"),
            "AZURE_DEPLOYMENT_NAME": os.getenv("AZURE_DEPLOYMENT_NAME"),
            "AZURE_SEARCH_ENDPOINT": os.getenv("AZURE_SEARCH_ENDPOINT"),
            "AZURE_SEARCH_INDEX_NAME": os.getenv("AZURE_SEARCH_INDEX_NAME"),
            "PF_DISABLE_TRACING": "true"
        }
    }
    
    deployment_file = deploy_dir / "deployment.yaml"
    with open(deployment_file, 'w') as f:
        yaml.dump(deployment_config, f, default_flow_style=False, sort_keys=False)
    print(f"✓ Created: {deployment_file}")
    
    # 3. Deployment commands script
    commands = f"""# Outlander Copilot Deployment Commands

## Prerequisites
- Azure CLI installed: https://aka.ms/azure-cli
- Logged in to Azure: `az login`
- ML extension installed: `az extension add -n ml`

## Configuration
$SUBSCRIPTION_ID = "{os.getenv('AZURE_SUBSCRIPTION_ID')}"
$RESOURCE_GROUP = "{os.getenv('AZURE_RESOURCE_GROUP')}"
$WORKSPACE = "{os.getenv('AZURE_AI_HUB_NAME', 'outlander-hub')}"
$LOCATION = "{os.getenv('AZURE_LOCATION', 'eastus2')}"

## Step 1: Create Managed Online Endpoint
az ml online-endpoint create `
  --file deployment/endpoint.yaml `
  --resource-group $RESOURCE_GROUP `
  --workspace-name $WORKSPACE `
  --subscription $SUBSCRIPTION_ID

## Step 2: Create Deployment
az ml online-deployment create `
  --file deployment/deployment.yaml `
  --resource-group $RESOURCE_GROUP `
  --workspace-name $WORKSPACE `
  --subscription $SUBSCRIPTION_ID `
  --all-traffic

## Step 3: Verify Deployment
az ml online-endpoint show `
  --name outlander-copilot-endpoint `
  --resource-group $RESOURCE_GROUP `
  --workspace-name $WORKSPACE

## Step 4: Get Endpoint Keys
az ml online-endpoint get-credentials `
  --name outlander-copilot-endpoint `
  --resource-group $RESOURCE_GROUP `
  --workspace-name $WORKSPACE

## Step 5: Test Deployment
$ENDPOINT_URI = az ml online-endpoint show `
  --name outlander-copilot-endpoint `
  --resource-group $RESOURCE_GROUP `
  --workspace-name $WORKSPACE `
  --query scoring_uri -o tsv

$API_KEY = az ml online-endpoint get-credentials `
  --name outlander-copilot-endpoint `
  --resource-group $RESOURCE_GROUP `
  --workspace-name $WORKSPACE `
  --query primaryKey -o tsv

# Test with curl or PowerShell
$headers = @{{
    "Content-Type" = "application/json"
    "Authorization" = "Bearer $API_KEY"
}}

$body = @{{
    chat_input = "Which tent is the most waterproof?"
    chat_history = @()
}} | ConvertTo-Json

Invoke-RestMethod -Uri $ENDPOINT_URI -Method Post -Headers $headers -Body $body

## Clean Up (After Testing)
# Delete deployment to avoid charges
az ml online-endpoint delete `
  --name outlander-copilot-endpoint `
  --resource-group $RESOURCE_GROUP `
  --workspace-name $WORKSPACE `
  --yes `
  --no-wait
"""
    
    commands_file = deploy_dir / "deploy_commands.ps1"
    with open(commands_file, 'w') as f:
        f.write(commands)
    print(f"✓ Created: {commands_file}")
    
    # 4. Deployment documentation
    docs = f"""# Outlander Copilot - Deployment Documentation

## Overview
This document describes the deployment configuration for the Outlander Copilot AI assistant.

## Architecture
- **Flow Type**: Prompt Flow (RAG-based chatbot)
- **Deployment Target**: Azure Machine Learning Managed Online Endpoint
- **Instance Type**: Standard_DS3_v2 (4 cores, 14GB RAM)
- **Scaling**: 1 instance (can be scaled up)

## Dependencies
- Azure OpenAI (GPT-4o): {os.getenv('AZURE_OPENAI_ENDPOINT')}
- Azure AI Search: {os.getenv('AZURE_SEARCH_ENDPOINT')}
- Search Index: {os.getenv('AZURE_SEARCH_INDEX_NAME')}

## Environment Variables
The deployment requires the following environment variables:
- `AZURE_OPENAI_ENDPOINT`: OpenAI service endpoint
- `AZURE_OPENAI_API_VERSION`: API version
- `AZURE_DEPLOYMENT_NAME`: GPT-4o deployment name
- `AZURE_SEARCH_ENDPOINT`: AI Search service endpoint
- `AZURE_SEARCH_INDEX_NAME`: Product search index name

## Deployment Files
- `endpoint.yaml`: Endpoint configuration
- `deployment.yaml`: Deployment configuration
- `deploy_commands.ps1`: Deployment script

## Estimated Costs
- **Compute**: ~$0.19/hour (~$140/month for 24/7)
- **Storage**: Minimal (<$1/month)
- **OpenAI**: Pay-per-use (based on tokens)
- **AI Search**: Free tier (already provisioned)

## Deployment Steps
See `deploy_commands.ps1` for detailed commands.

## Testing
Once deployed, test the endpoint using:
```powershell
Invoke-RestMethod -Uri $ENDPOINT_URI -Method Post -Headers $headers -Body $body
```

## Monitoring
- View logs: Azure Portal → ML Workspace → Endpoints
- Check metrics: Request rate, latency, errors
- Monitor costs: Cost Management + Billing

## Rollback
To rollback to a previous version:
```powershell
az ml online-endpoint update `
  --name outlander-copilot-endpoint `
  --traffic "outlander-deployment-v1=0,outlander-deployment-v2=100"
```

## Cleanup
To avoid ongoing costs, delete the endpoint after testing:
```powershell
az ml online-endpoint delete --name outlander-copilot-endpoint --yes
```

## Deployment Status
- Configuration files: ✓ Ready
- Azure resources: ✓ Provisioned
- Flow validation: ✓ Passed
- Local testing: ✓ Successful
- Ready for deployment: ✓ Yes

## Notes
For Udacity submission, deploy for testing and capture screenshots, then delete to avoid charges.
"""
    
    docs_file = deploy_dir / "DEPLOYMENT_README.md"
    with open(docs_file, 'w', encoding='utf-8') as f:
        f.write(docs)
    print(f"✓ Created: {docs_file}")
    
    print("\n" + "=" * 80)
    print("DEPLOYMENT CONFIGURATION FILES CREATED")
    print("=" * 80)
    print(f"\nFiles created in: {deploy_dir}")
    print("\n1. endpoint.yaml - Endpoint configuration")
    print("2. deployment.yaml - Deployment configuration")
    print("3. deploy_commands.ps1 - PowerShell deployment script")
    print("4. DEPLOYMENT_README.md - Documentation")
    print("\nNext steps:")
    print("1. Review the configuration files")
    print("2. Run: cd deployment; .\\deploy_commands.ps1")
    print("3. Test the deployed endpoint")
    print("4. Capture screenshots for Udacity submission")
    print("5. Delete the endpoint to avoid charges")
    print("=" * 80)


if __name__ == "__main__":
    create_deployment_configs()
