requirements.txt
# Azure AI and OpenAI SDKs
azure-ai-inference==1.0.0b5
azure-identity==1.19.0
openai==1.55.3
azure-search-documents==11.6.0b6

# Azure Computer Vision and AI Services
azure-ai-vision-imageanalysis==1.0.0b3
azure-cognitiveservices-vision-computervision==0.9.0

# Prompt Flow SDK
promptflow==1.15.0
promptflow-tools==1.4.0
promptflow-azure==1.15.0

# Azure Core Libraries
azure-core==1.31.0
azure-ai-ml==1.22.1

# Data Processing
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.1.3

# Image Processing and Computer Vision
Pillow==11.0.0
opencv-python==4.10.0.84
matplotlib==3.9.2

# Evaluation and Testing
pytest==8.3.3
pytest-asyncio==0.24.0

# HTTP and API
requests==2.32.3
aiohttp==3.11.7

# Utilities
pyyaml==6.0.2
jsonlines==4.0.0
tqdm==4.67.0

# Type Hints
typing-extensions==4.12.2

# Logging and Monitoring
python-json-logger==2.0.7
